Stop throwing for jet phi == NaN
Some jets seem to have a phi value equal to NaN: I'm not sure where
this came from, but it was breaking an assert statement (later
promoted to an exception throw). It seems to be very rare, so rather
than throwing the code was rewritten such that the NaN values just
propagate through the calculations.