Clean up scripts to make plots from hdf5 files
Made the hdf5 "state of the art" plotter able to figure out which
files are signal or background by default.

This also involved some refactoring of the object that gets cross
sections: it now lives in its own file and reads the cross section
file in the repository by default.